refactored test_admin_integration.py (#1662)
* refactored test_admin_integration.py

* Fixed the test issue after refactoring test_admin_integration

* fixed the imports and renamed test_admin_integration to mixins.py

* Update change log and added name to authors and fixed the formatting issue

* Update change log and added name to authors and fixed the formatting issue

* Updated requested changes

---------
