+++ AUTHORS
@@ -145,3 +145,4 @@ The following is a list of much appreciated contributors:
 * bgelov (Oleg Belov)
 * EricOuma (Eric Ouma)
 * Glay00 (Gleb Gorelov)
+* PrashansaChaudhary (Prashansa Chaudhary)
